test : `Repository.existsByName` 메소드 테스트 추가 및 `StationAcceptanceStep` 분리
